add gates in user to control access done.

+++ backend/app/Http/Controllers/UserController.php
@@ -16,12 +16,15 @@ class UserController extends Controller
 
     public function index()
     {
+        $this->authorize('view','users');
         return UserResource::collection(User::with('role')->paginate());
     }
 
 
     public function store(UserCreateRequest $request)
     {
+        $this->authorize('edit','users');
+
         $user = User::create(
             $request->only('first_name', 'last_name','email', 'role_id')
             + ['password' => Hash::make(1234)]
@@ -32,12 +35,15 @@ class UserController extends Controller
 
     public function show(string $id)
     {
+        $this->authorize('view','users');
         return new UserResource(User::with('role')->find($id));
     }
 
 
     public function update(UserUpdateRequest $request, string $id)
     {
+        $this->authorize('edit','users');
+
         $user = User::find($id);
 
         $user->update($request->only('first_name', 'last_name', 'email', 'role_id'));
@@ -48,6 +54,7 @@ class UserController extends Controller
 
     public function destroy(string $id)
     {
+        $this->authorize('edit','users');
         User::destroy($id);
 
         return response(null,Response::HTTP_NO_CONTENT);
